updated main

+++ main.c
@@ -11,7 +11,7 @@ int main(int argc, char **argv)
 
     gtk_window_set_title(GTK_WINDOW(window), "Vk Incognito");
     gtk_window_set_resizable(GTK_WINDOW(window), gtk_true());
-    gtk_window_resize(GTK_WINDOW(window), gint(800), gint(600));
+    gtk_window_resize(GTK_WINDOW(window), 800, 600);
     g_signal_connect(G_OBJECT(window), "destroy", G_CALLBACK(gtk_main_quit), NULL);
 
     gtk_widget_show(window);
